+++ .idea/dictionaries/zoid23.xml
@@ -1,0 +1,3 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="zoid23" />
+</component>
