+++ androidstories-app/src/main/res/layout/fragment_story.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".story.StoryDetailsFragment">
+    tools:context=".storydetails.StoryDetailsFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
